fix: V4 polish - streaming layer pull + dtype= + helpers + requests dep
Revisao linha-por-linha encontrou:
A) torch_dtype= deprecated em transformers 4.46+ -> dtype=.
B) blob = r.content carregava layer inteira (538MB) em RAM = pico 1GB+
   por layer. Streaming via tempfile + iter_content(8MB).
C) timeout layer pull 600s -> 1200s (margem rede oscilante Kaggle).
D) stream_layer com 5 niveis nesting -> 3 helpers (_download_blob,
   _apply_whiteout, _extract_member).
E) import io removido (nao usado mais com tempfile streaming).
F) requirements.txt: requests>=2.31.0 declarado.

+++ eval/_common.py
@@ -48,9 +48,9 @@ def load_model(adapter, base=BASE_MODEL):
     log.info(f"[load_model] device={device} dtype={dtype} base={base}")
     tokenizer = AutoTokenizer.from_pretrained(base)
     log.info("[load_model] tokenizer loaded, loading base model...")
-    base_model = AutoModelForCausalLM.from_pretrained(
-        base, torch_dtype=dtype, low_cpu_mem_usage=True
-    ).to(device)
+    base_model = AutoModelForCausalLM.from_pretrained(base, dtype=dtype, low_cpu_mem_usage=True).to(
+        device
+    )
     log.info(f"[load_model] base model on {device}, adapter={adapter}")
     if adapter and (Path(adapter) / "adapter_config.json").exists():
         from peft import PeftModel
